FIN DU RUSTLING

+++ exercises/23_conversions/using_as.rs
@@ -9,12 +9,10 @@
 //
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
-
-// I AM NOT DONE
-
+/// ajout de as f64 qui est donc son types
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64 
 }
 
 fn main() {
